+++ packages/apollo-tools/jest.config.js
@@ -1,0 +1,6 @@
+const baseConfig = require("../../jest.base");
+
+module.exports = {
+  ...baseConfig,
+  setupFiles: ["apollo-env"]
+};