+++ packages/apollo/__mocks__/apollo-cli-test.ts
@@ -26,8 +26,13 @@ const debug = fn => {
 };
 
 const deleteFolderRecursive = path => {
-  // don't relete files on azure CI
-  if (process.env.AZURE_HTTP_USER_AGENT) return;
+  // don't delete files on windows -- will get a resource locked error
+  if (
+    require("os")
+      .type()
+      .includes("Windows")
+  )
+    return;
 
   if (fs.existsSync(path)) {
     fs.readdirSync(path).forEach(function(file, index) {