+++ packages/vscode-apollo/src/languageServerClient.ts
@@ -54,7 +54,7 @@ export function getLanguageServerClient(
     ],
     synchronize: {
       fileEvents: [
-        workspace.createFileSystemWatcher("**/.env"),
+        workspace.createFileSystemWatcher("**/.env?(.local)"),
         workspace.createFileSystemWatcher(
           "**/*.{graphql,js,ts,jsx,tsx,vue,py,rb,dart,re}"
         )