+++ CHANGELOG.md
@@ -23,6 +23,50 @@
 - `vscode-apollo`
   - <First `vscode-apollo` related entry goes here>
 
+## `apollo@2.21.1`
+
+- `apollo@2.21.1`
+  - Add sdl download ability to `client:download-schema` [#1470](https://github.com/apollographql/apollo-tooling/pull/1470)
+  - colors: use cyan instead of blue for text highlighting [#1598](https://github.com/apollographql/apollo-tooling/pull/1598)
+  - Fix codegen --watch breaking out of watch mode on validation errors [#1627](https://github.com/apollographql/apollo-tooling/pull/1627)
+- `apollo-env@0.6.0`
+  - POTENTIALLY BREAKING: Make `apollo-env` a standard TS package
+    [#1611](https://github.com/apollographql/apollo-tooling/pull/1611) This PR likely warrants a pre-major version bump so that it isn't accidentally upgraded to for dependents using the ^ range. If this breaks your project, please don't hesitate to let us know and revert back to v0.5.1. This PR removes the handwritten node-fetch types and instead
+    just re-exports them. Internally, this drastically simplifies the existence of apollo-env within this monorepo, as it no longer requires a special build step that forces lerna and TS to work around it.
+- `vscode-apollo@1.12.1`
+  - Fix client localSchemaFile for vscode stats command [#1634](https://github.com/apollographql/apollo-tooling/pull/1634)
+
+## `apollo@2.21.0`
+
+- `apollo@2.21.0`
+  - Add flag to omit deprecated enum cases for swift codegen [#1595](https://github.com/apollographql/apollo-tooling/pull/1595)
+- `apollo-codegen-swift@0.35.7`
+  - Add flag to omit deprecated enum cases [#1595](https://github.com/apollographql/apollo-tooling/pull/1595)
+  - Fix code generation for input fields with the value `null` [#1596](https://github.com/apollographql/apollo-tooling/pull/1596)
+
+## `apollo@2.20.0`
+
+- `apollo@2.20.0`
+  - Fix codegen --watch mode not writing changes for files [#1591](https://github.com/apollographql/apollo-tooling/pull/1591)
+  - Fix `service:check` not outputing url
+- `apollo-codegen-swift@0.35.11`
+  - Fix code generation for empty input objects / arrays [#1589](https://github.com/apollographql/apollo-tooling/pull/1589)
+- `apollo-language-server@1.17.0`
+  - Improve autocomplete performance by fixing `leading` option and debounce threshold [#1593](https://github.com/apollographql/apollo-tooling/pull/1593)
+  - Load ENV variables from both `.env` and `.env.local` for Vue CLI compatibility [#1525](https://github.com/apollographql/apollo-tooling/pull/1525)
+- `vscode-apollo@1.12.0`
+  - Fix `ctrl+space` autocomplete with language-server imrovements [#1593](https://github.com/apollographql/apollo-tooling/pull/1593)
+  - Watch both `.env` and `.env.local` for changes [#1525](https://github.com/apollographql/apollo-tooling/pull/1525)
+
+## `apollo@2.19.1`
+
+- `apollo@2.19.1`
+  - Support custom frontend and historic parameters for federated service:check
+- `apollo-codegen-swift@0.35.10`
+  - Ensure fields named `self` don't cause compilation errors in the generated code [#1533](https://github.com/apollographql/apollo-tooling/pull/1533)
+  - Preserve leading/trailing underscores on field names [#1533](https://github.com/apollographql/apollo-tooling/pull/1533)
+  - Allow generated code to be compiled without a module umbrella header [#1248](https://github.com/apollographql/apollo-tooling/pull/1248)
+
 ## `apollo@2.19.0`
 
 - `apollo@2.19.0`